+++ stellaris-modding-support/src/main/gen/com/windea/plugin/idea/stellaris/localization/psi/StellarisLocalizationLocale.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 import com.intellij.openapi.util.Iconable.IconFlags;
+import com.intellij.psi.PsiReference;
 import com.windea.plugin.idea.stellaris.domain.StellarisLocale;
 import javax.swing.Icon;
 
@@ -27,5 +28,8 @@ public interface StellarisLocalizationLocale extends StellarisLocalizationNamedE
 
   @Nullable
   StellarisLocale getLocale();
+
+  @NotNull
+  PsiReference getReference();
 
 }