+++ stellaris-modding-support/src/main/gen/com/windea/plugin/idea/stellaris/localization/psi/impl/StellarisLocalizationIconImpl.java
@@ -9,6 +9,7 @@ import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.windea.plugin.idea.stellaris.localization.psi.StellarisLocalizationTypes.*;
 import com.windea.plugin.idea.stellaris.localization.psi.*;
+import com.intellij.psi.PsiReference;
 
 public class StellarisLocalizationIconImpl extends StellarisLocalizationNamedElementImpl implements StellarisLocalizationIcon {
 
@@ -52,6 +53,12 @@ public class StellarisLocalizationIconImpl extends StellarisLocalizationNamedEle
   @Override
   public int getTextOffset() {
     return StellarisLocalizationPsiImplUtil.getTextOffset(this);
+  }
+
+  @Override
+  @NotNull
+  public PsiReference getReference() {
+    return StellarisLocalizationPsiImplUtil.getReference(this);
   }
 
 }