+++ stellaris-modding-support/src/main/gen/com/windea/plugin/idea/stellaris/localization/psi/impl/StellarisLocalizationLocaleImpl.java
@@ -10,6 +10,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static com.windea.plugin.idea.stellaris.localization.psi.StellarisLocalizationTypes.*;
 import com.windea.plugin.idea.stellaris.localization.psi.*;
 import com.intellij.openapi.util.Iconable.IconFlags;
+import com.intellij.psi.PsiReference;
 import com.windea.plugin.idea.stellaris.domain.StellarisLocale;
 import javax.swing.Icon;
 
@@ -62,6 +63,12 @@ public class StellarisLocalizationLocaleImpl extends StellarisLocalizationNamedE
   @Nullable
   public StellarisLocale getLocale() {
     return StellarisLocalizationPsiImplUtil.getLocale(this);
+  }
+
+  @Override
+  @NotNull
+  public PsiReference getReference() {
+    return StellarisLocalizationPsiImplUtil.getReference(this);
   }
 
 }