+++ stellaris-modding-support/src/main/kotlin/com/windea/plugin/idea/stellaris/localization/reference/StellarisLocalizationPropertyPsiReference.kt
@@ -2,6 +2,7 @@
 
 package com.windea.plugin.idea.stellaris.localization.reference
 
+import com.intellij.codeInsight.lookup.*
 import com.intellij.model.*
 import com.intellij.openapi.util.*
 import com.intellij.psi.*
@@ -29,7 +30,7 @@ class StellarisLocalizationPropertyPsiReference(
 	//注意要传入elementName而非element
 	override fun getVariants(): Array<out Any> {
 		return findAllLocalizationPropertiesInProject(element.project).mapArray {
-			createLookupElement(it.name!!, icon = it.getIcon(0), typeText = it.containingFile.name)
+			LookupElementBuilder.create(it.name!!).withIcon(it.getIcon(0)).withTypeText(it.containingFile.name)
 		}
 	}
 }